Updated digit formatting for the 7segments.
Python script can convert the "font" to a format which minimizes bit shifting for the 7segment. 14 and 16 segment displays are not yet converted.

+++ display_protocol_testing/bmw_obc_display_lib.h
@@ -1,3 +1,5 @@
+#include "stdint.h"
+
 const uint8_t SevenSegmentASCII[96] = {
 	0b00000000, /* (space) */
 	0b10000110, /* ! */
@@ -196,7 +198,7 @@ const uint16_t FourteenSegmentASCII[96] = {
 	0b000000000000000, /* (del) */
 };
 
-const uint16_t SixteenSegmentASCII[96] = {
+const uint32_t SixteenSegmentASCII[96] = {
 	0b00000000000000000, /* (space) */
 	0b10000000000001100, /* ! */
 	0b00000001000000100, /* " */
